Add logo paths to useSite and move helper
