+++ src/common/utils/urlHelpers.js
@@ -1,13 +1,4 @@
-import {
-  AUDIO,
-  VIDEO,
-  IMAGE,
-  ORIGINAL,
-  SMALL,
-  MEDIUM,
-  THUMBNAIL,
-} from 'common/constants'
-
+// IMPORTANT getMediaUrl is now deprecated V1_FUDGE
 // size only applies to images
 // Values for size can be: 'Thumbnail', 'Small', 'Medium', 'FullHD', 'OriginalJpeg',
 export const getMediaUrl = ({ id, type, viewName = 'Medium' }) => {
@@ -29,27 +20,6 @@ export const getMediaUrl = ({ id, type, viewName = 'Medium' }) => {
 
     default:
       return 'The media type supplied is not recognised by the url helper'
-  }
-}
-
-// NB ALL sizes supplied for VIDEO or images of mime-type 'gif' will return a static image src except for ORIGINAL
-export const getMediaPath = ({ mediaObject, type, size = ORIGINAL }) => {
-  if (!mediaObject?.original) {
-    return "A media object with the property of 'originl' must be supplied to retrieve a src path."
-  }
-  if (![ORIGINAL, SMALL, MEDIUM, THUMBNAIL].includes(size)) {
-    return 'Only ORIGINAL, SMALL, MEDIUM, or THUMBNAIL are accepted as sizes for media.'
-  }
-  switch (type) {
-    case AUDIO:
-      return mediaObject?.original?.path
-
-    case VIDEO:
-    case IMAGE:
-      return mediaObject?.[size]?.path
-
-    default:
-      return 'The media type supplied is not recognised by the getMediaPath helper'
   }
 }
