+++ src/context/SiteContext.js
@@ -10,7 +10,9 @@ export const siteInitialState = {
     title: '',
     sitename: '',
     parentLanguageTitle: '',
-    logoId: '',
+    logoPathMedium: '',
+    logoPathSmall: '',
+    logoPathThumbnail: '',
     topBackgroundImageId: '',
     topBackgroundVideoId: '',
     joinText: null,
